+++ src/modules/resources/components/index.ts
@@ -1,1 +1,2 @@
 export { DocumentOutline } from './DocumentOutline';
+export { FirecrawlSettings } from './FirecrawlSettings';