+++ src/modules/resources/url/scrapers/index.ts
@@ -1,0 +1,2 @@
+export { FirecrawlScraper } from './FirecrawlScraper';
+export { BasicFetchScraper } from './BasicFetchScraper';